+++ plugins/source/aws-pricing/client/client.go
@@ -3,6 +3,7 @@ package client
 import (
 	"context"
 	"fmt"
+
 	"github.com/cloudquery/plugin-sdk/plugins/source"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/specs"
@@ -10,12 +11,12 @@ import (
 )
 
 type Client struct {
-	Logger zerolog.Logger
+	Logger   zerolog.Logger
+	Endpoint string
 }
 
 func (c *Client) ID() string {
-	// TODO: Change to either your plugin name or a unique dynamic identifier
-	return "ID"
+	return "awspricing"
 }
 
 func New(ctx context.Context, logger zerolog.Logger, s specs.Source, opts source.Options) (schema.ClientMeta, error) {
@@ -24,9 +25,10 @@ func New(ctx context.Context, logger zerolog.Logger, s specs.Source, opts source
 	if err := s.UnmarshalSpec(&pluginSpec); err != nil {
 		return nil, fmt.Errorf("failed to unmarshal plugin spec: %w", err)
 	}
-	// TODO: Add your client initialization here
+	pluginSpec.SetDefaults()
 
 	return &Client{
-		Logger: logger,
+		Logger:   logger,
+		Endpoint: pluginSpec.Endpoint,
 	}, nil
 }
