+++ plugins/source/aws-pricing/resources/products.go
@@ -1,0 +1,23 @@
+package resources
+
+import (
+	"context"
+
+	"github.com/cloudquery/plugin-sdk/schema"
+	"github.com/cloudquery/plugin-sdk/transformers"
+)
+
+func products() *schema.Table {
+	return &schema.Table{
+		Name:      "awspricing_service_products",
+		Resolver:  fetchProducts,
+		Transform: transformers.TransformWithStruct(&Product{}, transformers.WithPrimaryKeys("Sku")),
+		Relations: []*schema.Table{},
+	}
+}
+
+func fetchProducts(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- interface{}) error {
+	pricingFile := parent.Item.(PricingFile)
+	res <- pricingFile.Products
+	return nil
+}
